+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Semver.
 - AVEDEV average absolute deviation
 - MODE.SNGL most frequent value (tie-broken by first occurrence)
 - PERCENTILE.INC, PERCENTILE.EXC, QUARTILE.INC, QUARTILE.EXC
+- LARGE, SMALL k-th largest/smallest value
 
 ## [0.2.1] - 2026-05-11